pyproject.toml
[project]
name = "bankgpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "playwright",
    "llama-cloud-services",
    "llama-index-core",
    "llama-index-readers-file",
    "python-dotenv",
    "llama_index.llms.gemini",
    "llama_index.embeddings.google_genai",
    "llama_index.llms.google_genai",
    "llama_index.vector_stores.postgres",
    "python-dotenv",
    "mcp[cli,core]>=1.6.0",
    "chainlit",
    "anthropic",
    "asyncpg",
    "browser-use>=0.4.0",
    "mistralai",
    "supabase",
    "fastmcp>=2.10.1",
    "psycopg2-binary",
    "sqlalchemy",
    "dspy-ai>=2.6.0",
]
